assembler: Move CSR definitions into another header
With all the CSRs added, this is quite a lot to scan past before getting
to the actual assembler interface.

This is also a little more convenient to users as well when using the
CSR instruction generators, since "Assembler::" doesn't need to be
prefixed on the CSR enumeration accesses, which makes for less writing.